% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor_hooks.R
\name{check_description_json}
\alias{check_description_json}
\title{Check for description JSON corresponding to a SNIRF file}
\usage{
check_description_json(snirf_path)
}
\arguments{
\item{snirf_path}{path to a SNIRF (.snirf) file}
}
\value{
path to the description JSON if found
}
\description{
Check for description JSON corresponding to a SNIRF file
}
