% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetDescription.R
\name{datasetDescription_server}
\alias{datasetDescription_server}
\title{Dataset Description Shiny Module Server}
\usage{
datasetDescription_server(id, converted_root)
}
\arguments{
\item{id}{Character. Shiny module namespace ID.}

\item{converted_root}{Reactive. The path to the folder where the BIDS dataset will be saved.}
}
\value{
A list containing:
  \describe{
    \item{dataset_name}{Reactive expression returning the dataset name input by the user.}
  }
}
\description{
This Shiny module server handles the dataset description inputs from the user
and saves a BIDS-compliant `dataset_description.json` file.
}
