% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetDescription.R
\name{datasetDescription_ui}
\alias{datasetDescription_ui}
\title{Helper function for dataset description}
\usage{
datasetDescription_ui(id)
}
\arguments{
\item{id}{takes app id}
}
\description{
Helper function for dataset description
}
