% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentalDesign.R
\name{experimentalDesign_server}
\alias{experimentalDesign_server}
\title{Helper function for experimentalDesign_server}
\usage{
experimentalDesign_server(id, converted_root, dataset_name_reactive)
}
\arguments{
\item{id}{takes app id}

\item{dataset_name_reactive}{takes dataset}

\item{currentConvertedPathReactive}{takes current path}
}
\description{
Helper function for experimentalDesign_server
}
