% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentalDesign.R
\name{experimentalDesign_ui}
\alias{experimentalDesign_ui}
\title{Helper function for experimentalDesign}
\usage{
experimentalDesign_ui(id)
}
\arguments{
\item{id}{takes app id}
}
\description{
Helper function for experimentalDesign
}
