% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readme.R
\name{Readme_server}
\alias{Readme_server}
\title{Readme Shiny Module Server}
\usage{
Readme_server(id, converted_root)
}
\arguments{
\item{id}{Character. Shiny module namespace ID.}

\item{converted_root}{Reactive. The path to the folder where the BIDS dataset will be saved.}
}
\description{
Readme Shiny Module Server
}
