% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskMapping.R
\name{taskMapping_server}
\alias{taskMapping_server}
\title{Helper function for task mapping server}
\usage{
taskMapping_server(id, converted_root, routine)
}
\arguments{
\item{id}{takes app id}

\item{routine}{either "json" or "folders"}
}
\description{
Helper function for task mapping server
}
