% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentalDesign.R
\name{experimentalDesign_server}
\alias{experimentalDesign_server}
\title{Helper function for experimentalDesign_server}
\usage{
experimentalDesign_server(
  id,
  currentConvertedPathReactive,
  dataset_name_reactive
)
}
\arguments{
\item{id}{takes app id}

\item{currentConvertedPathReactive}{takes current path}

\item{dataset_name_reactive}{takes dataset}
}
\description{
Helper function for experimentalDesign_server
}
