% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{snirf2bids}
\alias{snirf2bids}
\title{snirf2bids function for conversion}
\usage{
snirf2bids(source_snirf, converted_root, experiment_description)
}
\arguments{
\item{experiment}{description}
}
\description{
snirf2bids function for conversion
}
