% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskMapping.R
\name{taskMapping_server}
\alias{taskMapping_server}
\title{Helper function for task mapping server}
\usage{
taskMapping_server(id, dataset_name_reactive)
}
\arguments{
\item{id}{takes app id}

\item{dataset_name_reactive}{takes dataset name}
}
\description{
Helper function for task mapping server
}
